rename feature to result-preview
https://coveord.atlassian.net/browse/KIT-530
