--- packages/headless/src/features/quickview/quickview-state.ts
@@ -1,11 +1,0 @@
-export interface QuickviewState {
-  resultUniqueId: string;
-  resultContent: string;
-}
-
-export function getQuickviewInitialState(): QuickviewState {
-  return {
-    resultUniqueId: '',
-    resultContent: '',
-  };
-}
